`@remotion/browser-studio`: Save composition `defaultProps` to source
