+++ packages/studio-codemods/src/index.ts
@@ -46,6 +46,12 @@ export {JsxElementNotFoundAtLocationError} from './sequence-props/jsx-element-no
 export {simpleDiff} from './simple-diff';
 export {splitJsxSequence} from './split-jsx-sequence';
 export {splitVideoFromAudio} from './split-video-from-audio';
+export {
+	formatInlineContentWithFormatter,
+	getCompositionDefaultPropsLine,
+	updateDefaultProps,
+	type FormatInline,
+} from './update-default-props';
 export {updateInlineCaptionPatches} from './update-inline-caption-patches';
 export {
 	type RemovedProp,
