+++ packages/studio-server/src/codemods/format-inline-content.ts
@@ -1,3 +1,5 @@
+import {formatInlineContentWithFormatter} from '@remotion/studio-codemods';
+
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
 type PrettierType = typeof import('prettier');
 
@@ -9,10 +11,6 @@ type PrettierType = typeof import('prettier');
  * @param linePrefix - Everything from the start of the line to where
  *   inlineContent will appear (used to calculate column offset and indentation)
  * @param endOfLine - Prettier endOfLine option
- *
- * We wrap the content in `const __x__ = CONTENT;` and adjust printWidth
- * so prettier makes the same line-breaking decisions as if the content
- * were at its actual column position in the file.
  */
 export const formatInlineContent = async ({
 	inlineContent,
@@ -50,61 +48,11 @@ export const formatInlineContent = async ({
 		}
 	}
 
-	const tabWidth = (prettierConfig.tabWidth as number) ?? 2;
-	const baseIndent = linePrefix.match(/^(\s*)/)?.[1] ?? '';
-
-	// Calculate visual column offset (tabs expand to tabWidth columns)
-	const columnOffset = [...linePrefix].reduce(
-		(col, ch) => (ch === '\t' ? col + tabWidth : col + 1),
-		0,
-	);
-
-	// Adjust printWidth so the wrapper prefix occupies the same visual
-	// width as the actual file prefix, ensuring identical line breaks.
-	const configPrintWidth = (prettierConfig.printWidth as number) ?? 80;
-	const wrapperPrefix = 'const __x__ = ';
-	const effectivePrintWidth = Math.max(
-		configPrintWidth - columnOffset + wrapperPrefix.length,
-		20,
-	);
-
-	const wrappedSource = `${wrapperPrefix}${inlineContent};\n`;
-	const formattedWrapped = await format(wrappedSource, {
-		...prettierConfig,
-		printWidth: effectivePrintWidth,
-		filepath: 'test.tsx',
-		plugins: [],
+	return formatInlineContentWithFormatter({
+		inlineContent,
+		linePrefix,
 		endOfLine,
+		prettierConfig,
+		format: (source, options) => format(source, {...options, plugins: []}),
 	});
-
-	// Extract the formatted value from the wrapper
-	const withoutSemicolon = formattedWrapped.replace(/;\s*$/, '');
-	const wrappedInParentheses = withoutSemicolon.startsWith(
-		`${wrapperPrefix}(\n`,
-	);
-	let formattedProps: string;
-
-	if (withoutSemicolon.startsWith(wrapperPrefix) && !wrappedInParentheses) {
-		formattedProps = withoutSemicolon.slice(wrapperPrefix.length);
-	} else {
-		// Prettier broke the line after `=` — extract and dedent one level
-		const lines = withoutSemicolon
-			.split('\n')
-			.slice(1, wrappedInParentheses ? -1 : undefined);
-		const useTabs = prettierConfig.useTabs as boolean;
-		const oneIndent = useTabs ? '\t' : ' '.repeat(tabWidth);
-		formattedProps = lines
-			.map((l) => (l.startsWith(oneIndent) ? l.slice(oneIndent.length) : l))
-			.join('\n');
-	}
-
-	// Add base indentation to all lines except the first
-	const indentedProps = formattedProps
-		.split('\n')
-		.map((line, i) =>
-			i === 0 ? line : line.length > 0 ? baseIndent + line : line,
-		)
-		.join('\n');
-
-	return {formatted: indentedProps, didFormat: true};
 };
