+++ packages/studio/src/components/RenderQueue/actions.ts
@@ -395,7 +395,7 @@ export const callUpdateDefaultPropsApi = (
 	defaultProps: Record<string, unknown>,
 	enumPaths: EnumPath[],
 ) => {
-	return callApi('/api/update-default-props', {
+	const body = {
 		compositionId,
 		defaultProps: NoReactInternals.serializeJSONWithSpecialTypes({
 			data: defaultProps,
@@ -403,7 +403,13 @@ export const callUpdateDefaultPropsApi = (
 			staticBase: window.remotion_staticBase,
 		}).serializedString,
 		enumPaths,
-	});
+	};
+	const browserStudioOperations = getBrowserStudioOperations();
+	if (browserStudioOperations !== null) {
+		return browserStudioOperations.updateDefaultProps(body);
+	}
+
+	return callApi('/api/update-default-props', body);
 };
 
 export const applyVisualControlChange = ({
